Extract network
